Add mobile menu icon
